Create shared mood utilities
