+++ client/src/pages/DashBoardHome.jsx
@@ -10,6 +10,7 @@ import PlantGrid from '../components/PlantGrid';
 import { checkAndGrowPlant } from '../services/plantService';
 import ProgressRing from '../components/ProgressRing';
 import { getWeeklyEngagement } from '../utils/engagement.js';
+import { calculateMoodStreak } from '../utils/moodUtils';
 import RecommendationBanner from '../components/RecommendationBanner.jsx';
 import './DashboardHome.css';
 
@@ -76,59 +77,6 @@ function DashBoardHome() {
     }
   }, [displayName]);
 
-  // Helper function to calculate streak from mood logs
-  const calculateStreak = (logs) => {
-    if (!logs || logs.length === 0) return 0;
-
-    // Convert dates to YYYY-MM-DD strings
-    const toDayString = (date) => {
-      return new Date(date).toISOString().split('T')[0];
-    };
-
-    // Get all unique dates with logs
-    const dateSet = new Set();
-    logs.forEach(log => {
-      dateSet.add(toDayString(log.createdAt));
-    });
-
-    // Convert to array and sort (newest first)
-    const uniqueDates = Array.from(dateSet).sort().reverse();
-
-    // Check if there's a log for today
-    const today = toDayString(new Date());
-
-    // Find the most recent log date
-    let mostRecentLogDate = uniqueDates[0];
-
-    // If the most recent log is not from today or yesterday, streak is broken
-    if (mostRecentLogDate !== today) {
-      const yesterday = toDayString(new Date(Date.now() - 86400000));
-      if (mostRecentLogDate !== yesterday) {
-        return 0;
-      }
-    }
-
-    // Start counting streak from most recent log
-    let currentStreak = 1;
-    let currentDate = mostRecentLogDate;
-
-    // Check for consecutive days working backward
-    while (true) {
-      // Get the previous day
-      const prevDate = toDayString(new Date(new Date(currentDate).getTime() - 86400000));
-
-      // If there's a log for the previous day, increment streak
-      if (dateSet.has(prevDate)) {
-        currentStreak++;
-        currentDate = prevDate;
-      } else {
-        // Gap found, streak ends
-        break;
-      }
-    }
-
-    return currentStreak;
-  };
 
   // Fetch mood logs and calculate weekly logs count and streak
   // This will run on initial load and whenever todayMood changes
@@ -156,11 +104,10 @@ function DashBoardHome() {
         setWeeklyLogs(weeklyLogsSet.size);
 
         // Calculate streak using only mood logs
-        const currentStreak = calculateStreak(moodLogs);
+        const currentStreak = calculateMoodStreak(moodLogs);
         setStreak(currentStreak);
 
       } catch (error) {
-        console.error('Error fetching logs:', error);
         toast.error('Unable to load logs.');
         setWeeklyLogs(0);
         setStreak(0);
